Исправление бага, при котором невозможно было что-либо удалять, если у него отсутсвует изображение.

+++ app/classes/Propel/Models/Comment.php
@@ -59,11 +59,14 @@ class Comment extends BaseComment
 	 */
 	public function preDelete(ConnectionInterface $connection = null)
 	{
-		if (is_file(\Fenric\Upload::path($this->getPicture())))
+		if ($this->getPicture())
 		{
-			if (is_readable(\Fenric\Upload::path($this->getPicture())))
+			if (is_file(\Fenric\Upload::path($this->getPicture())))
 			{
-				unlink(\Fenric\Upload::path($this->getPicture()));
+				if (is_readable(\Fenric\Upload::path($this->getPicture())))
+				{
+					unlink(\Fenric\Upload::path($this->getPicture()));
+				}
 			}
 		}
