+++ app/classes/Propel/Models/Section.php
@@ -85,11 +85,14 @@ class Section extends BaseSection
 	 */
 	public function preDelete(ConnectionInterface $connection = null)
 	{
-		if (is_file(\Fenric\Upload::path($this->getPicture())))
+		if ($this->getPicture())
 		{
-			if (is_readable(\Fenric\Upload::path($this->getPicture())))
+			if (is_file(\Fenric\Upload::path($this->getPicture())))
 			{
-				unlink(\Fenric\Upload::path($this->getPicture()));
+				if (is_readable(\Fenric\Upload::path($this->getPicture())))
+				{
+					unlink(\Fenric\Upload::path($this->getPicture()));
+				}
 			}
 		}
